+++ package.nix
@@ -8,7 +8,7 @@
 }:
 rustPlatform.buildRustPackage {
   pname = "skillmill";
-  version = "0.1.2";
+  version = "0.1.3";
 
   src = ./.;
   cargoLock = {